refactor: 파일명, 폴더 구조, DOM 구조 등 변경
- rename
  - @util/auth.js -> @util/fetchProfile.js
  - @components/ProvideAuth.jsx -> @components/ProvideAuth 폴더로 이동
- ProvideAuth로 App을 감싸는 부분이 index.js파일으로 이동
