+++ src/frontend/components/ProvideAuth/index.js
@@ -1,0 +1,1 @@
+export * from './ProvideAuth.jsx';
